refactor: Вынести страницы в собственные директории
