+++ src/components/app/app.module.css
@@ -13,7 +13,3 @@
   overflow: hidden;
   width: 1240px;
 }
-
-.message {
-  align-self: center;
-}
