--- src/components/tab/tab.module.css
@@ -1,7 +1,0 @@
-.link {
-  text-decoration: none;
-}
-
-.tab {
-  height: 64px;
-}
