+++ src/pages/constructor/constructor.jsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import { useSelector } from 'react-redux';
-import { selectIngredientsSlice } from '../utils/selectors';
-import styles from '../components/app/app.module.css';
+import { selectIngredientsSlice } from '../../utils/selectors';
+import styles from './constructor.module.css';
 import { DndProvider } from 'react-dnd';
 import { HTML5Backend } from 'react-dnd-html5-backend';
-import BurgerIngredients from '../components/burger-ingredients/burger-ingredients';
-import BurgerConstructor from '../components/burger-constructor/burger-constructor';
+import BurgerIngredients from '../../components/burger-ingredients/burger-ingredients';
+import BurgerConstructor from '../../components/burger-constructor/burger-constructor';
 
 function ConstructorPage() {
   const ingredients = useSelector(selectIngredientsSlice);
