+++ src/pages/constructor/constructor.module.css
@@ -1,0 +1,3 @@
+.message {
+  align-self: center;
+}
