+++ src/pages/index.js
@@ -1,10 +1,10 @@
-import ConstructorPage from './constructor';
-import LoginPage from './login';
-import RegistrationPage from './registration';
-import ForgotPasswordPage from './forgot-password';
-import ResetPasswordPage from './reset-password';
-import ProfilePage from './profile';
-import IngredientPage from './ingredient';
+import ConstructorPage from './constructor/constructor';
+import LoginPage from './login/login';
+import RegistrationPage from './registration/registration';
+import ForgotPasswordPage from './forgot-password/forgot-password';
+import ResetPasswordPage from './reset-password/reset-password';
+import ProfilePage from './profile/profile';
+import IngredientPage from './ingredient/ingredient';
 export {
   ConstructorPage,
   LoginPage,
