+++ src/pages/profile/profile.module.css
@@ -1,6 +1,6 @@
 .links {
-  margin: 0;
   padding: 0;
+  margin: 0;
   list-style: none;
 }
 
@@ -15,6 +15,16 @@
 
 .buttons {
   display: flex;
-  align-items: center;
   justify-content: center;
+  align-items: center;
+}
+
+.links {
+  padding: 0;
+  margin: 0;
+  list-style: none;
+}
+
+.link {
+  text-decoration: none;
 }
