+++ src/pages/reset-password/reset-password.jsx
@@ -1,11 +1,11 @@
 import React, { useState } from 'react';
-import styles from './page.module.css';
+import styles from '../page.module.css';
 import { Button, Input, PasswordInput } from '@ya.praktikum/react-developer-burger-ui-components';
 import { Link, Redirect, useHistory } from 'react-router-dom';
-import { RouterPaths } from '../utils/constants';
-import { resetPasswordRequest } from '../utils/burger-api';
+import { RouterPaths } from '../../utils/constants';
+import { resetPasswordRequest } from '../../utils/burger-api';
 import { useSelector } from 'react-redux';
-import { selectIsUserAuthorized } from '../utils/selectors';
+import { selectIsUserAuthorized } from '../../utils/selectors';
 
 function ResetPasswordPage() {
   const [form, setValue] = useState({ password: '', token: '' });
@@ -29,7 +29,7 @@ function ResetPasswordPage() {
   }
 
   return (
-    <div className={styles.container}>
+    <div className={`mt-30 ${styles.container}`}>
       <form className={`mb-20`}>
         <h1 className={'text text_type_main-medium'}>Восстановление пароля</h1>
         <PasswordInput
@@ -52,7 +52,7 @@ function ResetPasswordPage() {
       </form>
       <p className={'text text_type_main-default text_color_inactive'}>
         Вспомнили пароль?{' '}
-        <Link to={RouterPaths.LOGIN} className={`${styles.link} colors-interface-accent`}>
+        <Link to={RouterPaths.LOGIN} className={`${styles.link} text_color_accent`}>
           Войти
         </Link>
       </p>
